requirements.txt
# --- Core Backend ---
fastapi[standard]
uvicorn[standard]
pydantic
python-multipart

# --- OCR & Document Processing ---
ocrmypdf
pytesseract
pdfplumber
#pymupdf==1.24.5
python-docx
openpyxl
pillow

# --- Local LLM ---
llama-cpp-python

# --- XML & Validation ---
lxml
xmlschema

# --- PDF Generation ---
reportlab

# --- Utilities ---
python-dateutil
loguru
